refactor(core): move local runtime implementations out of deps/services
deps/services.py had grown into a runtime-implementation module: the
accepted-note preparer factory, current-note content freshener, note-file
delete storage/enqueuer, directory-delete relation cleanup refresher,
the four background schedulers with their task machinery, and the
project-index route command all lived inside the FastAPI composition
root. Move them into the index package so the composition root only
wires dependencies:

- index/local_notes.py: accepted-note mutation and note-file cleanup
  implementations (LocalAcceptedNotePreparerFactory,
  LocalAcceptedNoteRepositories, LocalCurrentNoteContentFreshener and
  its protocols, LocalNoteFileDeleteStorage,
  LocalDirectoryFileDeleteEnqueuer,
  LocalDirectoryDeleteRelationCleanupRefresher)
- index/local_schedulers.py: background-task machinery
  (drain_background_tasks and friends) plus the four local schedulers
  and their capability protocols
- index/local_project.py: project-index route protocols
  (ProjectIndexRunner/Observer/Scheduler/Command),
  ProjectIndexRouteRequest, and LocalProjectIndexCommand, next to the
  LocalProjectIndexRunner they orchestrate

deps/services.py keeps every provider and Dep annotation and re-exports
drain_background_tasks for the API/MCP/CLI lifespans, so those call
sites are untouched. ProjectIndexRouteRequest is no longer exported from
basic_memory.deps; the route and tests import it from
basic_memory.index.local_project. Scheduler and accepted-note tests move
to tests/index alongside the code.
