+++ src/basic_memory/deps/__init__.py
@@ -101,7 +101,6 @@ from basic_memory.deps.services import (
     ProjectIndexSchedulerDep,
     get_project_index_command,
     ProjectIndexCommandDep,
-    ProjectIndexRouteRequest,
     get_search_reindex_scheduler,
     SearchReindexSchedulerDep,
     get_search_service,
@@ -258,7 +257,6 @@ __all__ = [
     "ProjectIndexSchedulerDep",
     "get_project_index_command",
     "ProjectIndexCommandDep",
-    "ProjectIndexRouteRequest",
     "get_search_reindex_scheduler",
     "SearchReindexSchedulerDep",
     "get_search_service",
