+++ src/basic_memory/deps/services.py
@@ -1,24 +1,22 @@
 """Service dependency injection for basic-memory.
 
-This module provides service-layer dependencies:
+This module is the FastAPI composition root for the service layer: every
+provider constructs services from repositories, config, and the local runtime
+implementations that live in ``basic_memory.index`` — it defines no runtime
+behavior of its own:
 - EntityParser, MarkdownProcessor
 - FileService, EntityService
 - SearchService, LinkResolver, ContextService
 - ProjectService, DirectoryService
+- local note-content, indexing, and background-scheduler runtimes
 """
 
-import asyncio
-from collections.abc import Sequence
-from dataclasses import dataclass
 from pathlib import Path
-from typing import Annotated, Any, Coroutine, Protocol
+from typing import Annotated
 
 from fastapi import Depends
 from loguru import logger
-from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
-from basic_memory import db
-from basic_memory.config import BasicMemoryConfig
 from basic_memory.deps.config import AppConfigDep
 from basic_memory.deps.db import SessionMakerDep
 from basic_memory.deps.projects import (
@@ -41,11 +39,7 @@ from basic_memory.deps.repositories import (
     SearchRepositoryV2Dep,
     SearchRepositoryV2ExternalDep,
 )
-from basic_memory.indexing.relation_resolution import (
-    RelationResolutionRuntime,
-    RepositoryRelationResolutionRuntime,
-    resolve_project_relations,
-)
+from basic_memory.indexing.relation_resolution import RepositoryRelationResolutionRuntime
 from basic_memory.cloud import (
     DirectoryDeleteService,
     LocalNoteContentMaterializationProvider,
@@ -53,40 +47,49 @@ from basic_memory.cloud import (
     NoteContentQueryService,
 )
 from basic_memory.index.local_dependencies import build_local_markdown_file_indexer
-from basic_memory.index.local_project import LocalProjectIndexObservation, LocalProjectIndexRunner
-from basic_memory.indexing.project_index_coordinator import ProjectIndexCoordinatorResult
+from basic_memory.index.local_notes import (
+    LocalAcceptedNotePreparerFactory,
+    LocalAcceptedNoteRepositories,
+    LocalCurrentNoteContentFreshener,
+    LocalDirectoryDeleteRelationCleanupRefresher,
+    LocalDirectoryFileDeleteEnqueuer,
+)
+from basic_memory.index.local_project import (
+    LocalProjectIndexCommand,
+    LocalProjectIndexRunner,
+    ProjectIndexCommand,
+    ProjectIndexObserver,
+    ProjectIndexRunner,
+    ProjectIndexScheduler,
+)
+
+# Re-exported lifecycle hook: the API, MCP, and one-shot CLI lifespans drain
+# scheduled background work through the composition root before shutting down
+# the database.
+from basic_memory.index.local_schedulers import drain_background_tasks as drain_background_tasks
+from basic_memory.index.local_schedulers import (
+    EntityVectorSyncScheduler,
+    LocalEntityVectorSyncScheduler,
+    LocalProjectIndexScheduler,
+    LocalRelationResolutionScheduler,
+    LocalSearchReindexScheduler,
+    RelationResolutionScheduler,
+    SearchReindexScheduler,
+)
 from basic_memory.indexing.accepted_note_mutation_runner import (
     AcceptedNoteMutationDependencies,
     AcceptedNoteMutationMovePolicy,
-    AcceptedNoteMutationPreparer,
 )
 from basic_memory.indexing.batch_indexer import BatchIndexer
 from basic_memory.indexing.directory_delete_runner import (
     DirectoryDeleteRuntime,
-    DirectoryFileDeleteEnqueueError,
     RepositoryDirectoryDeleteAcceptanceStore,
 )
 from basic_memory.indexing.index_file_runner import IndexFileExecutor
 from basic_memory.indexing.models import StorageIndexFileWriter
-from basic_memory.indexing.note_file_delete_runner import run_note_file_delete
-from basic_memory.file_utils import FileError
 from basic_memory.markdown import EntityParser
 from basic_memory.markdown.markdown_processor import MarkdownProcessor
-from basic_memory.models import Project
-from basic_memory.repository import NoteContentRepository, ObservationRepository, RelationRepository
-from basic_memory.repository.accepted_note_search_repository import AcceptedNoteSearchRepository
-from basic_memory.repository.entity_repository import EntityRepository
-from basic_memory.repository.search_repository import create_search_repository
-from basic_memory.runtime.cleanup import RuntimeFileDeleteResult, RuntimeNoteFileDeleteJobRequest
-from basic_memory.runtime.storage import (
-    ProjectId,
-    RuntimeFileChecksum,
-    RuntimeFilePath,
-    runtime_content_type_is_markdown,
-)
-from basic_memory.schemas import ProjectIndexRunResponse
 from basic_memory.services import EntityService, ProjectService
-from basic_memory.services.exceptions import FileOperationError
 from basic_memory.services.context_service import ContextService
 from basic_memory.services.directory_service import DirectoryService
 from basic_memory.services.file_service import FileService
@@ -256,190 +259,7 @@ NoteContentQueryServiceDep = Annotated[
 ]
 
 
-@dataclass(frozen=True, slots=True)
-class LocalAcceptedNotePreparerFactory:
-    """Construct prepare-only note semantics for local accepted-note mutations."""
-
-    session_maker: async_sessionmaker[AsyncSession]
-    app_config: BasicMemoryConfig
-
-    def create_note_preparer(self, project: Project) -> AcceptedNoteMutationPreparer:
-        entity_parser = EntityParser(Path(project.path))
-        markdown_processor = MarkdownProcessor(entity_parser, app_config=self.app_config)
-        file_service = FileService(
-            Path(project.path),
-            markdown_processor,
-            app_config=self.app_config,
-        )
-        entity_repository = EntityRepository(project_id=project.id)
-        search_repository = create_search_repository(
-            self.session_maker,
-            project_id=project.id,
-            app_config=self.app_config,
-        )
-        search_service = SearchService(
-            search_repository,
-            entity_repository,
-            file_service,
-            self.session_maker,
-        )
-        link_resolver = LinkResolver(
-            entity_repository=entity_repository,
-            search_service=search_service,
-            session_maker=self.session_maker,
-        )
-        return EntityService(
-            entity_repository=entity_repository,
-            observation_repository=ObservationRepository(project_id=project.id),
-            relation_repository=RelationRepository(project_id=project.id),
-            entity_parser=entity_parser,
-            file_service=file_service,
-            link_resolver=link_resolver,
-            session_maker=self.session_maker,
-            search_service=search_service,
-            app_config=self.app_config,
-        )
-
-
-class LocalCurrentNoteEntity(Protocol):
-    """Entity fields needed to refresh current markdown before route mutation."""
-
-    file_path: str
-    content_type: str
-
-
-class LocalCurrentNoteEntityRepository(Protocol):
-    """Entity lookup needed by the local note-content freshener."""
-
-    async def get_by_external_id(
-        self,
-        session: AsyncSession,
-        external_id: str,
-        *,
-        load_relations: bool = True,
-    ) -> LocalCurrentNoteEntity | None: ...
-
-
-class LocalCurrentNoteFileService(Protocol):
-    """Current file-state access needed before mutating accepted note content."""
-
-    async def exists(
-        self,
-        path: RuntimeFilePath,
-    ) -> bool: ...
-
-
-class LocalCurrentNoteFileIndexer(Protocol):
-    """Single-file indexing capability used by the local note-content freshener."""
-
-    async def index_file(
-        self,
-        file_path: RuntimeFilePath,
-        *,
-        source: str,
-    ) -> object: ...
-
-
-@dataclass(frozen=True, slots=True)
-class LocalCurrentNoteContentFreshener:
-    """Converge directly-edited local markdown before accepted-note mutations."""
-
-    entity_repository: LocalCurrentNoteEntityRepository
-    file_service: LocalCurrentNoteFileService
-    file_indexer: LocalCurrentNoteFileIndexer
-    session_maker: async_sessionmaker[AsyncSession]
-
-    async def freshen_note_content(
-        self,
-        *,
-        project_external_id: str,
-        entity_external_id: str,
-    ) -> None:
-        del project_external_id
-
-        async with self.session_maker() as session:
-            entity = await self.entity_repository.get_by_external_id(
-                session,
-                entity_external_id,
-                load_relations=False,
-            )
-            if entity is None or not runtime_content_type_is_markdown(entity):
-                return
-            file_path = entity.file_path
-
-        if not await self.file_service.exists(file_path):
-            return
-
-        await self.file_indexer.index_file(
-            file_path,
-            source="note-content-mutation-freshen",
-        )
-
-
 # --- Directory Delete Runtime ---
-
-
-@dataclass(frozen=True, slots=True)
-class LocalNoteFileDeleteStorage:
-    """Adapt local FileService to guarded materialized-note cleanup."""
-
-    file_service: FileService
-
-    async def exists(self, path: RuntimeFilePath) -> bool:
-        return await self.file_service.exists(path)
-
-    async def compute_checksum(self, path: RuntimeFilePath) -> RuntimeFileChecksum:
-        return await self.file_service.compute_checksum(path)
-
-    async def delete_file(self, path: RuntimeFilePath) -> None:
-        await self.file_service.delete_file(path)
-
-
-@dataclass(frozen=True, slots=True)
-class LocalDirectoryFileDeleteEnqueuer:
-    """Run accepted directory-delete file cleanup inline for the local runtime."""
-
-    file_service: FileService
-
-    async def enqueue_directory_file_delete(
-        self,
-        request: RuntimeNoteFileDeleteJobRequest,
-    ) -> RuntimeFileDeleteResult | None:
-        # Local cleanup runs inline, so return the guarded delete result; a skipped
-        # delete (file changed before cleanup) must not be reported as a success.
-        try:
-            return await run_note_file_delete(
-                request,
-                storage=LocalNoteFileDeleteStorage(file_service=self.file_service),
-            )
-        except (FileError, FileOperationError, OSError) as exc:
-            raise DirectoryFileDeleteEnqueueError(str(exc)) from exc
-
-
-@dataclass(frozen=True, slots=True)
-class LocalDirectoryDeleteRelationCleanupRefresher:
-    """Reindex surviving relation sources after an accepted directory delete.
-
-    Their search_index relation rows went stale when the deleted targets'
-    rows cascaded away; reindexing each surviving source drops the danglers.
-    """
-
-    session_maker: async_sessionmaker[AsyncSession]
-    entity_repository: EntityRepository
-    search_service: SearchService
-
-    async def refresh_relation_sources(self, entity_ids: Sequence[int]) -> None:
-        unique_entity_ids = sorted(set(entity_ids))
-        if not unique_entity_ids:
-            return
-
-        async with db.scoped_session(self.session_maker) as session:
-            entities = await self.entity_repository.find_by_ids(session, unique_entity_ids)
-
-        # A source deleted between acceptance and this refresh has no search rows
-        # left to repair, so missing ids are skipped rather than treated as fatal.
-        for entity in entities:
-            await self.search_service.index_entity(entity)
 
 
 async def get_directory_delete_service(
@@ -707,28 +527,6 @@ IndexFileExecutorV2ExternalDep = Annotated[
 # --- Note Content Writes ---
 
 
-@dataclass(frozen=True, slots=True)
-class LocalAcceptedNoteRepositories:
-    """Project-scoped core repositories for accepted-note mutations.
-
-    One concrete bundle satisfies both the lookup and write repository
-    capabilities the accepted-note mutation runner consumes; cloud composes its
-    own tenant-scoped equivalent behind the same protocols.
-    """
-
-    def entity_repository(self, project_id: ProjectId) -> EntityRepository:
-        return EntityRepository(project_id=project_id)
-
-    def pending_entity_repository(self, project_id: ProjectId) -> EntityRepository:
-        return EntityRepository(project_id=project_id)
-
-    def note_content_repository(self, project_id: ProjectId) -> NoteContentRepository:
-        return NoteContentRepository(project_id=project_id)
-
-    def search_repository(self, project_id: ProjectId) -> AcceptedNoteSearchRepository:
-        return AcceptedNoteSearchRepository(project_id=project_id)
-
-
 async def get_note_content_mutation_service(
     project_repository: ProjectRepositoryDep,
     entity_repository: EntityRepositoryV2ExternalDep,
@@ -775,51 +573,6 @@ NoteContentMutationServiceDep = Annotated[
 # --- Project Indexing ---
 
 
-class ProjectIndexRunner(Protocol):
-    """Run project-wide indexing in the current process."""
-
-    async def index_project(
-        self,
-        project_id: int,
-        *,
-        force_full: bool = False,
-    ) -> ProjectIndexCoordinatorResult: ...
-
-
-class ProjectIndexObserver(Protocol):
-    """Observe project files visible to the active runtime."""
-
-    async def observe_project(self, project_id: int) -> LocalProjectIndexObservation: ...
-
-
-class ProjectIndexScheduler(Protocol):
-    """Schedule background project indexing."""
-
-    def schedule_project_index(self, *, project_id: int, force_full: bool = False) -> None: ...
-
-
-@dataclass(frozen=True, slots=True)
-class ProjectIndexRouteRequest:
-    """Route-level project-index command input."""
-
-    project_id: int
-    project_name: str
-    force_full: bool
-    run_in_background: bool
-
-
-type ProjectIndexRouteResult = ProjectIndexRunResponse | dict[str, str]
-
-
-class ProjectIndexCommand(Protocol):
-    """Handle a project-index route request."""
-
-    async def index_project(
-        self,
-        request: ProjectIndexRouteRequest,
-    ) -> ProjectIndexRouteResult: ...
-
-
 async def get_project_index_runner(
     project_repository: ProjectRepositoryDep,
     session_maker: SessionMakerDep,
@@ -852,162 +605,6 @@ ProjectIndexObserverDep = Annotated[
 # --- Background Work Schedulers ---
 
 
-class EntityVectorSyncScheduler(Protocol):
-    """Schedule out-of-band semantic vector refreshes for note mutations."""
-
-    def schedule_entity_vector_sync(self, *, entity_id: int, project_id: int) -> None: ...
-
-
-class SearchReindexScheduler(Protocol):
-    """Schedule a search-index rebuild for the active project."""
-
-    def schedule_search_reindex(self, *, project_id: int) -> None: ...
-
-
-class EntityVectorSyncSearchService(Protocol):
-    async def sync_entity_vectors(self, entity_id: int) -> object: ...
-
-
-class SearchReindexService(Protocol):
-    async def reindex_all(self) -> object: ...
-
-
-def _log_task_failure(completed: asyncio.Task) -> None:
-    if completed.cancelled():
-        return
-    try:
-        completed.result()
-    except asyncio.CancelledError:
-        return
-    except Exception as exc:  # pragma: no cover
-        logger.exception("Background task failed", error=str(exc))
-
-
-# The event loop holds only weak references to tasks; without a strong reference
-# a suspended background task can be garbage-collected mid-flight and silently
-# never finish (asyncio.create_task docs: "Save a reference to the result").
-_background_tasks: set[asyncio.Task[object]] = set()
-
-
-def _schedule_background_coroutine(
-    coroutine: Coroutine[Any, Any, object],
-    *,
-    test_mode: bool,
-) -> None:
-    # Background tasks outlive pytest fixture cleanup and can race engine disposal.
-    # Focused tests call the scheduler classes directly with test_mode=False.
-    if test_mode:
-        coroutine.close()
-        return
-
-    task = asyncio.create_task(coroutine)
-    _background_tasks.add(task)
-    task.add_done_callback(_background_tasks.discard)
-    task.add_done_callback(_log_task_failure)
-
-
-async def drain_background_tasks() -> None:
-    """Await scheduled background work until none remains.
-
-    One-shot CLI clients close the event loop right after the command coroutine
-    returns, which would cancel in-flight vector sync and relation resolution
-    scheduled by the write path — leaving semantic search stale until an
-    unrelated reindex. A task can schedule a follow-up task (the
-    relation-resolution dirty re-run), so drain in waves until no running task
-    remains. Failures are already logged by the done callback; the drain itself
-    never raises.
-    """
-    while True:
-        # Filter on task state, not set membership: completed tasks are pruned
-        # by a call_soon done-callback that may not have run yet, and awaiting
-        # only already-done tasks never suspends — checking membership alone
-        # would busy-spin without ever letting that callback fire.
-        running = [task for task in _background_tasks if not task.done()]
-        if not running:
-            break
-        await asyncio.wait(running)
-
-
-@dataclass(frozen=True, slots=True)
-class LocalEntityVectorSyncScheduler:
-    search_service: EntityVectorSyncSearchService
-    test_mode: bool
-
-    def schedule_entity_vector_sync(self, *, entity_id: int, project_id: int) -> None:
-        _ = project_id
-        _schedule_background_coroutine(
-            self.search_service.sync_entity_vectors(entity_id),
-            test_mode=self.test_mode,
-        )
-
-
-# Process-lifetime single-flight state: project ids with an index run already
-# scheduled or in flight. Every POST .../index and the startup scan previously
-# spawned an independent full coordinator run over the same rows; overlapping
-# runs are also the trigger for move/delete races, so at most one run per
-# project may be in flight.
-_pending_project_index: set[int] = set()
-# Projects whose index request arrived while a run was already in flight, with
-# the strongest force_full seen. The in-flight run scanned a snapshot that may
-# predate the new request, so exactly one trailing rerun starts when it
-# finishes — mirroring the relation-resolution dirty bit above.
-_dirty_project_index: dict[int, bool] = {}
-
-
-@dataclass(frozen=True, slots=True)
-class LocalProjectIndexScheduler:
-    """Run background project indexing with per-project single-flight coalescing."""
-
-    project_index_runner: ProjectIndexRunner
-    test_mode: bool
-
-    def schedule_project_index(self, *, project_id: int, force_full: bool = False) -> None:
-        # Early-return in test mode BEFORE touching the pending set: the
-        # background coroutine (which clears the set) never runs under test mode,
-        # so adding here would leak the project id forever.
-        if self.test_mode:
-            return
-        # Coalesce: a run is already pending/in flight for this project. Mark it
-        # dirty (keeping the strongest force_full) so one follow-up run covers
-        # this request once the current run finishes, instead of racing it.
-        if project_id in _pending_project_index:
-            _dirty_project_index[project_id] = (
-                _dirty_project_index.get(project_id, False) or force_full
-            )
-            return
-        _pending_project_index.add(project_id)
-        _schedule_background_coroutine(
-            self._run_project_index(project_id, force_full=force_full),
-            test_mode=self.test_mode,
-        )
-
-    async def _run_project_index(self, project_id: int, *, force_full: bool) -> None:
-        try:
-            await self.project_index_runner.index_project(project_id, force_full=force_full)
-        finally:
-            rerun_force_full = _dirty_project_index.pop(project_id, None)
-            _pending_project_index.discard(project_id)
-            # Re-arm inside finally, outside the in-flight window (pending now
-            # cleared), so a request that raced the run gets its own pass even
-            # when this run raised — a failed run is exactly when the coalesced
-            # request most needs its retry. Bounded to one extra run per burst.
-            if rerun_force_full is not None:
-                self.schedule_project_index(project_id=project_id, force_full=rerun_force_full)
-
-
-@dataclass(frozen=True, slots=True)
-class LocalSearchReindexScheduler:
-    search_service: SearchReindexService
-    test_mode: bool
-
-    def schedule_search_reindex(self, *, project_id: int) -> None:
-        _ = project_id
-        _schedule_background_coroutine(
-            self.search_service.reindex_all(),
-            test_mode=self.test_mode,
-        )
-
-
 async def get_entity_vector_sync_scheduler(
     search_service: SearchServiceV2ExternalDep,
     app_config: AppConfigDep,
@@ -1036,79 +633,6 @@ async def get_search_reindex_scheduler(
         search_service=search_service,
         test_mode=app_config.is_test_env,
     )
-
-
-class RelationResolutionScheduler(Protocol):
-    """Schedule background forward-reference resolution after note mutations."""
-
-    def schedule_relation_resolution(self, *, project_id: int) -> None: ...
-
-
-# Process-lifetime coalescing state: project ids with a relation-resolution
-# pass already pending or in flight. A burst of writes collapses to a single
-# offline pass instead of one whole-project relation scan per write — running a
-# scan per write made the write path heavier and piled up under concurrency
-# (see benchmarks/docs/write-load-benchmark.md).
-_pending_relation_resolution: set[int] = set()
-# Project ids whose forward references arrived while a pass was already scanning.
-# The scan resolves whatever is unresolved when it reads the table, so a write that
-# commits during the scan (after that read) would otherwise be missed until an
-# unrelated later trigger. This dirty bit forces exactly one follow-up pass.
-_dirty_relation_resolution: set[int] = set()
-
-
-@dataclass(frozen=True, slots=True)
-class LocalRelationResolutionScheduler:
-    """Back-resolve dangling forward references off the request path, coalesced.
-
-    The MCP/API write path inline-indexes the materialized note but never
-    back-resolves inbound `[[wikilinks]]` whose target the new note now
-    satisfies (#1015). Resolution is a whole-project scan, so running it per
-    write is both wasteful and a real write-load cost. Instead each write only
-    enqueues: the first write of a burst schedules one debounced background pass
-    and every other write coalesces onto it (at most one pending pass per
-    project). The accept path stays light; reconciliation runs offline. No-op in
-    test mode, consistent with the other local schedulers.
-    """
-
-    relation_runtime: RelationResolutionRuntime
-    test_mode: bool
-    debounce_seconds: float = 0.5
-
-    def schedule_relation_resolution(self, *, project_id: int) -> None:
-        # Early-return in test mode BEFORE touching the pending set: the
-        # background coroutine (which clears the set) never runs under test mode,
-        # so adding here would leak the project id forever.
-        if self.test_mode:
-            return
-        # Coalesce: a pass is already pending/running for this project. Mark it
-        # dirty so a scan that has already read the table re-runs once more and
-        # picks up this write's rows, instead of dropping it (#1002 review).
-        if project_id in _pending_relation_resolution:
-            _dirty_relation_resolution.add(project_id)
-            return
-        _pending_relation_resolution.add(project_id)
-        _schedule_background_coroutine(
-            self._resolve_after_debounce(project_id),
-            test_mode=self.test_mode,
-        )
-
-    async def _resolve_after_debounce(self, project_id: int) -> None:
-        try:
-            # Debounce: let the burst settle so one pass covers all of it.
-            await asyncio.sleep(self.debounce_seconds)
-            # Writes up to here are covered by the scan we are about to run, so only
-            # writes that land DURING the scan should force a re-run.
-            _dirty_relation_resolution.discard(project_id)
-            await resolve_project_relations(self.relation_runtime)
-        finally:
-            rerun = project_id in _dirty_relation_resolution
-            _dirty_relation_resolution.discard(project_id)
-            _pending_relation_resolution.discard(project_id)
-        # Re-arm outside the in-flight window (pending now cleared) so a write that
-        # raced the scan gets its own pass. Bounded to one extra pass per burst.
-        if rerun:
-            self.schedule_relation_resolution(project_id=project_id)
 
 
 async def get_relation_resolution_scheduler(
@@ -1187,41 +711,6 @@ NoteContentMaterializationProviderDep = Annotated[
     LocalNoteContentMaterializationProvider,
     Depends(get_note_content_materialization_provider),
 ]
-
-
-@dataclass(frozen=True, slots=True)
-class LocalProjectIndexCommand:
-    project_index_runner: ProjectIndexRunner
-    project_index_scheduler: ProjectIndexScheduler
-
-    async def index_project(
-        self,
-        request: ProjectIndexRouteRequest,
-    ) -> ProjectIndexRouteResult:
-        if request.run_in_background:
-            self.project_index_scheduler.schedule_project_index(
-                project_id=request.project_id,
-                force_full=request.force_full,
-            )
-            logger.info(
-                f"Filesystem indexing initiated for project: {request.project_name} "
-                f"(force_full={request.force_full})"
-            )
-
-            return {
-                "status": "index_started",
-                "message": (f"Filesystem indexing initiated for project '{request.project_name}'"),
-            }
-
-        result = await self.project_index_runner.index_project(
-            request.project_id,
-            force_full=request.force_full,
-        )
-        logger.info(
-            f"Filesystem indexing completed for project: {request.project_name} "
-            f"(force_full={request.force_full})"
-        )
-        return ProjectIndexRunResponse.from_result(result)
 
 
 async def get_project_index_command(
